hackflix/requirements.txt
APScheduler==3.2.0
click==6.7
cycler==0.10.0
decorator==4.3.0
Flask==0.11.1
Flask-Assets==0.12
httplib2==0.10.3
imageio==2.2.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
networkx==2.1
numpy==1.14.0
opencv-python==3.4.1.15
Pillow==5.2.0
psycopg2==2.7.1
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
qrcode==6.0
qrtools==0.0.2
requests==2.7.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.9
tzlocal==1.5.1
webassets==0.12.1
Werkzeug==0.14.1
zbar-py==1.0.4
